Disable negcache
When a DNS query fails (NXDOMAIN), dnsmasq will cache the response and
keep on replying with NXDOMAINs...

However, the domain may have been created/propagated in between (as a
system administrator, it happens to me quite often).

This commit disables caching of NXDOMAIN answers: if the domain is
created it will be immediately available locally.